requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
torch
git+https://github.com/openai/CLIP.git
